FEAT: Implement visual grid setup for Pac-Man game
Update method `setup_grid()` in the `PacManGame` class that utili-
zes a multiline string representation to define the game's maze.
This visual approach allows for easier editing and understanding
of the grid's layout, enhancing the maintainability and scalability
of the maze configuration.
It includes designations for paths, walls, power pellets, and tunn-
els, reflecting a classic Pac-Man style maze.

Author: Nicolas Rojas Bernal